+++ client/src/chartsProvider.ts
@@ -1,6 +1,6 @@
 import { join } from "path";
 import {
-    Event, EventEmitter, TextDocument, TextDocumentContentProvider, Uri, workspace,
+    TextDocument, Uri, ViewColumn, window, workspace
 } from "vscode";
 import { languageId } from "./extension";
 
@@ -10,72 +10,76 @@ export interface IConnectionDetails {
     url: string;
 }
 
-export class AxibaseChartsProvider implements TextDocumentContentProvider {
-    /**
-     * Sets the working document and fires update() event
-     */
-    public set document(document: TextDocument) {
-        this.innerDocument = document;
-        this.update();
-    }
-
-    public get document(): TextDocument {
-        return this.innerDocument;
-    }
-
-    public get onDidChange(): Event<Uri> {
-        return this.onDidChangeEmitter.event;
-    }
-
-    public static readonly previewUri: Uri = Uri.parse("axibaseCharts://authority/axibaseCharts");
+/**
+ * Stores connection details and extension context, creates WebviewPanel and sets it's content.
+ */
+export class ChartsProvider {
     private readonly absolutePath: (path: string) => string;
-
     private atsd!: boolean;
-    private innerDocument: TextDocument;
     private jsessionid: string | undefined;
-    private readonly onDidChangeEmitter: EventEmitter<Uri>;
     private text: string | undefined;
     private url!: string;
 
-    public constructor(details: IConnectionDetails, document: TextDocument,
-                       absolutePath: (path: string) => string) {
-        if (document.languageId !== languageId) {
-            throw new Error("Incorrect language");
-        }
-        this.onDidChangeEmitter = new EventEmitter<Uri>();
-        this.innerDocument = document;
-        this.updateSettings(details);
+    public constructor(absolutePath: (path: string) => string) {
         this.absolutePath = absolutePath;
     }
 
-    /**
-     * Applies the new settings to the current preview
-     * @param details new settings
-     */
-    public changeSettings(details: IConnectionDetails): void {
-        this.updateSettings(details);
-        this.update();
+    public createPanel() {
+        const panel = window.createWebviewPanel(
+            languageId,
+            "Preview Portal",
+            ViewColumn.Two,
+            {
+                // Allow <script>s in the webview content.
+                enableScripts: true,
+                // Allow the access to resources only in extension's resources directory.
+                localResourceRoots: [this.getUri("resources")]
+            }
+        );
+        return panel;
     }
 
     /**
-     * Provides html code to preview the configuration
+     * Prepares html for WebviewPanel.
+     * @param document Document, which content must be used as config for portal.
      */
-    public provideTextDocumentContent(): string {
-        this.text = deleteComments(this.innerDocument.getText());
-        this.clearUrl();
-        this.replaceImports();
-
-        this.addUrl();
-        const html: string = this.getHtml();
-
-        return html;
+    public getWebviewContent(document?: TextDocument): string {
+        if (this.url == null) {
+            /**
+             * Unable to connect to ATSD after changes in settings, see onDidChangeConfiguration.
+             */
+            return `<h3>Unable to connect to ASTD</h3>
+<p>Check connection <a href="https://github.com/axibase/axibase-charts-vscode#live-preview"> settings</a>.</p>`;
+        }
+        if (document) {
+            this.text = deleteComments(document.getText());
+            this.clearUrl();
+            this.replaceImports();
+            this.addUrl();
+        }
+        return this.getHtml();
     }
 
     /**
-     * Fires onDidChange to inform all listeners
+     * Updates the provider state
+     * @param details new connection details
      */
-    public update(): void {
-        this.onDidChangeEmitter.fire(AxibaseChartsProvider.previewUri);
+    public updateSettings(details: IConnectionDetails): void {
+        if (!details) {
+            /**
+             * Unable to connect to ATSD after changes in settings, see onDidChangeConfiguration.
+             */
+            this.url = this.jsessionid = this.atsd = null;
+            return;
+        }
+        this.url = details.url;
+        if (details.cookie) {
+            this.jsessionid = details.cookie[0].split(";")[0]
+                .split("=")[1];
+        } else {
+            this.jsessionid = undefined;
+        }
+        this.atsd = details.atsd;
     }
 
     /**
@@ -112,13 +116,21 @@ ${this.text.substr(match.index + match[0].length + 1)}`;
             this.url = this.url.substr(0, match.index);
         }
     }
-
+    /**
+     * Returns path to local resource as Uri.
+     * @param resource
+     */
+    private getUri(resource: string): Uri {
+        return Uri.file(
+            this.absolutePath(join("client", resource))
+        );
+    }
     /**
      * Generates the path to a resource on the local filesystem
      * @param resource path to a resource
      */
     private extensionPath(resource: string): string {
-        return this.absolutePath(join("client", resource));
+        return this.getUri(resource).with({ scheme: "vscode-resource" }).toString();
     }
 
     /**
@@ -143,13 +155,25 @@ ${this.text.substr(match.index + match[0].length + 1)}`;
 	  .portalPage body {
 		padding: 0;
 		background: var(--vscode-editor-background);
-	  }
+      }
+
+      body.vscode-light {
+        color: black;
+      }
+
+      body.vscode-dark {
+        color: white;
+      }
+
+      body.vscode-high-contrast {
+        color: red;
+      }
 	</style>
 	<script>
 	    window.previewOptions = ${JSON.stringify({
             jsessionid: this.jsessionid,
             text: this.text,
-            url: this.url,
+            url: this.url
         })};
 	</script>
 	<script src="${this.resource("portal_init.js")}"></script>
@@ -171,7 +195,7 @@ ${this.text.substr(match.index + match[0].length + 1)}`;
     /**
      * Adds the ATSD URL to the import statements
      * For example, `import fred = fred.js` becomes
-     * `import fred = https://nur.axibase.com/portal/resource/scripts/fred.js`
+     * `import fred = https://axibase.com/portal/resource/scripts/fred.js`
      */
     private replaceImports(): void {
         if (!this.text) {
@@ -192,7 +216,7 @@ ${this.text.substr(match.index + match[0].length + 1)}`;
     }
 
     /**
-     * Generates the path to a resource.
+     * Generates the path to a resource at server.
      * @param resource name of a static resource
      */
     private resource(resource: string): string {
@@ -202,21 +226,6 @@ ${this.text.substr(match.index + match[0].length + 1)}`;
         const resourcePath: string = `${cssType ? cssPath : jsPath}/${resource}`;
 
         return `${resourcePath}${this.jsessionid ? `;jsessionid=${this.jsessionid}` : ""}`;
-    }
-
-    /**
-     * Updates the provider state
-     * @param details new connection details
-     */
-    private updateSettings(details: IConnectionDetails): void {
-        this.url = details.url;
-        if (details.cookie) {
-            this.jsessionid = details.cookie[0].split(";")[0]
-                .split("=")[1];
-        } else {
-            this.jsessionid = undefined;
-        }
-        this.atsd = details.atsd;
     }
 }