+++ client/src/extension.ts
@@ -17,17 +17,16 @@ import {
     Range,
     TextDocument,
     Uri,
-    ViewColumn,
+    WebviewPanel,
     window,
     workspace,
-    WorkspaceConfiguration,
+    WorkspaceConfiguration
 } from "vscode";
 import {
-    ForkOptions, LanguageClient, LanguageClientOptions, ServerOptions, TransportKind,
+    ForkOptions, LanguageClient, LanguageClientOptions, ServerOptions, TransportKind
 } from "vscode-languageclient";
-import { AxibaseChartsProvider, IConnectionDetails } from "./axibaseChartsProvider";
+import { ChartsProvider, IConnectionDetails } from "./chartsProvider";
 import { statusFamily, StatusFamily, userAgent } from "./util";
-
 const configSection: string = "axibaseCharts";
 export const languageId: string = "axibasecharts";
 const errorMessage: string = "Configure connection properties in VSCode > Preferences > Settings. Open Settings," +
@@ -57,7 +56,10 @@ export const activate: (context: ExtensionContext) => void = async (context: Ext
     // Options to control the language client
     const clientOptions: LanguageClientOptions = {
         // Register the server for plain text documents
-        documentSelector: [{ language: languageId }],
+        documentSelector: [{
+            language: languageId,
+            scheme: "file"
+        }],
         outputChannelName: "Axibase Charts",
         synchronize: {
             // Notify the server about file changes to ".clientrc files contain in the workspace
@@ -102,68 +104,74 @@ export const activate: (context: ExtensionContext) => void = async (context: Ext
     // Start the client. This will also launch the server
     client.start();
 
-    let provider: AxibaseChartsProvider | undefined;
-    context.subscriptions.push(workspace.onDidSaveTextDocument((document: TextDocument) => {
-        if (provider && provider.document.uri === document.uri) {
-            provider.update();
+    let provider: ChartsProvider | undefined;
+    context.subscriptions.push(workspace.onDidChangeTextDocument((event) => {
+        if (event.document.languageId === languageId && provider) {
+            panel.webview.html = provider.getWebviewContent(event.document);
         }
     }));
+
     context.subscriptions.push(window.onDidChangeActiveTextEditor(() => {
-        if (window.activeTextEditor && window.activeTextEditor.document.languageId === languageId && provider) {
-            provider.document = window.activeTextEditor.document;
+        if (window.activeTextEditor &&
+            window.activeTextEditor.document.languageId === languageId && provider) {
+            panel.webview.html = provider.getWebviewContent(window.activeTextEditor.document);
         }
     }));
+
+    let details: IConnectionDetails;
+    let panel: WebviewPanel;
     context.subscriptions.push(commands.registerCommand(`${languageId}.showPortal`, async (): Promise<void> => {
         if (!window.activeTextEditor) {
-            return Promise.resolve();
+            return;
         }
 
+        const document: TextDocument = window.activeTextEditor.document;
+        if (document.languageId !== languageId) {
+            return;
+        }
+        /**
+         * One provider and one panel for all configs, only panel.webview.html is changed.
+         */
         if (!provider) {
-            const document: TextDocument = window.activeTextEditor.document;
-            if (document.languageId !== languageId) {
-                return Promise.resolve();
-            }
-            let details: IConnectionDetails;
-            try {
-                details = await constructConnection();
-            } catch (err) {
-                window.showErrorMessage(err);
-
-                return Promise.resolve();
-            }
-
-            provider = new AxibaseChartsProvider(details, document, context.asAbsolutePath);
-
-            context.subscriptions.push(workspace.registerTextDocumentContentProvider("axibaseCharts", provider));
-            provider.update();
+            provider = new ChartsProvider(context.asAbsolutePath);
+            await setConnectionDetails();
+            panel = provider.createPanel();
+            panel.onDidDispose(
+                () => {
+                    provider = null;
+                },
+                null,
+                context.subscriptions
+            );
         }
-        const tabLabel = "Preview Portal";
-        commands.executeCommand("vscode.previewHtml", AxibaseChartsProvider.previewUri, ViewColumn.Two, tabLabel)
-            .then(() => {
-                provider.update();
-            });
+        panel.webview.html = provider.getWebviewContent(document);
     }));
+
     context.subscriptions.push(
         workspace.onDidChangeConfiguration(async (e: ConfigurationChangeEvent): Promise<void> => {
             if (e.affectsConfiguration(configSection) && provider) {
                 const answer: string | undefined =
                     await window.showInformationMessage("Current connection details were modified.", "Reload");
                 if (answer === "Reload") {
-                    let details: IConnectionDetails;
-                    try {
-                        details = await constructConnection();
-                    } catch (err) {
-                        window.showErrorMessage(err);
-
-                        return Promise.resolve();
-                    }
-                    provider.changeSettings(details);
-
-                    return Promise.resolve();
+                    details = null;
+                    await setConnectionDetails();
+                    panel.webview.html = provider.getWebviewContent();
                 }
             }
         }),
     );
+
+    async function setConnectionDetails() {
+        if (!details) {
+            try {
+                details = await constructConnection();
+            } catch (err) {
+                window.showErrorMessage(err);
+                details = null;
+            }
+        }
+        provider.updateSettings(details);
+    }
 };
 
 /**
@@ -178,35 +186,33 @@ export const deactivate: () => Thenable<void> = (): Thenable<void> => {
 };
 
 /**
- * Constructs connection details based on the extension configuration and an input box
+ * Constructs connection details based on the extension configuration and an input box:
+ * tries to connect to server and get cookies, triggers window with error message
+ * if any connection error has been occurred, otherwise triggers inform window.
  */
-const constructConnection: () => Promise<IConnectionDetails> = async (): Promise<IConnectionDetails> => {
+async function constructConnection(): Promise<IConnectionDetails> {
     const config: WorkspaceConfiguration = workspace.getConfiguration(configSection);
     const protocol: string | undefined = config.get("protocol");
     if (!protocol) {
-        return Promise.reject(errorMessage);
+        throw new Error(errorMessage);
     }
     const address: string | undefined = config.get("hostname");
     if (!address) {
-        return Promise.reject(errorMessage);
+        throw new Error(errorMessage);
     }
     const port: number | undefined = config.get("port");
     if (!port) {
-        return Promise.reject(errorMessage);
+        throw new Error(errorMessage);
     }
     const url: string = `${protocol}://${address}:${port}`;
     let password: string | undefined;
     const username: string | undefined = config.get("username");
     if (username) {
-        try {
-            password = await window.showInputBox({
-                ignoreFocusOut: true, password: true,
-                prompt: `Enter the password for user ${username} to connect ` +
-                    `to ${address}:${port}. Exit VSCode to terminate the session.`,
-            });
-        } catch (err) {
-            return Promise.reject(err as Error);
-        }
+        password = await window.showInputBox({
+            ignoreFocusOut: true, password: true,
+            prompt: `Enter the password for user ${username} to connect ` +
+                `to ${address}:${port}. Exit VSCode to terminate the session.`,
+        });
     }
 
     let cookie: string[] | undefined;
@@ -217,13 +223,13 @@ const constructConnection: () => Promise<IConnectionDetails> = async (): Promise
         const error: Error = err as Error;
         window.showErrorMessage(error.toString());
 
-        return Promise.reject(error);
+        throw error;
     }
-    window.showInformationMessage(`Connected to ${address}:${port} ${username ? `as ${username}` : ""}`);
+    window.showInformationMessage(`Connected to ${address}:${port} ${username && password ? `as ${username}` : ""}`);
     atsd = atsd === undefined ? true : atsd;
 
     return { url, cookie, atsd };
-};
+}
 
 /**
  * Gets the cookies for a user from the server and tests if it ATSD or not
@@ -287,7 +293,7 @@ function handleResponse(
     const family: StatusFamily = statusFamily(res.statusCode);
     if (family === StatusFamily.CLIENT_ERROR || family === StatusFamily.SERVER_ERROR) {
         if (res.statusCode === 401) {
-            return reject(new Error(`Login failed with status code ${res.statusCode}`));
+            return reject(new Error(`Login failed with status code ${res.statusCode} ${res.statusMessage}`));
         } else {
             return reject(new Error(`Unexpected Response Code ${res.statusCode}`));
         }